Do not recompile installed standard library files
When .hs and .lvm timestamps are equal and both are
read-only, assume they are installed and up to date.

This is a minimal change for the case where the
installation tool resets the time stamps to the same
value for all files. Installations that preserve
modification times already worked. Tools that write
the files in some order and do not reset time stamps
require more complex behavior, because it starts to
overlap with 'normal' permission problem for which
the compiler user does not expect any special cases.