Fix not qualifying and wrong module in decls
- Fixed setting module when generating functions for use in the same
module
- Fixed types not being qualified when looking them up in the type
inference output (findCoreType function)
- Fixed newtypes not being generated with qualified names
- Removed a few instances where both the qualified and non-qualified
versions were generated in the desugaring phase
- Fix using the qualified version of Monad when desugaring do-notation
- Re-added deadcode elimination
- When importing a type from core into the importenvironment the
function typeSchemeFromCore expected all of the foralls at the beginning
of the function, (a rank-1 type) but toCoreDecl can generate types with
nested forall's (a rank-N type) when generating code accessing class
members (in classFunctions). At the moment this is handled by collecting
all of the quantified variables and treating it like a rank-1 type. This
can possibly cause weird behaviour when multiple class member functions
are used at the same time on the same instance type.